[#54] refactor: 프로젝트 구조 통일
