+++ src/main/java/org/unilab/improfessorbe/domain/notice/service/NoticeService.java
@@ -8,9 +8,9 @@ import org.springframework.transaction.annotation.Transactional;
 import org.unilab.improfessorbe.domain.notice.domain.Notice;
 import org.unilab.improfessorbe.domain.notice.dto.request.NoticeRequest;
 import org.unilab.improfessorbe.domain.notice.dto.response.NoticeResponse;
-import org.unilab.improfessorbe.domain.notice.repository.NoticeRepository;
+import org.unilab.improfessorbe.domain.notice.infrastructure.repository.NoticeRepository;
 import org.unilab.improfessorbe.domain.user.domain.User;
-import org.unilab.improfessorbe.domain.user.repository.UserRepository;
+import org.unilab.improfessorbe.domain.user.infrastructure.repository.UserRepository;
 import org.unilab.improfessorbe.global.exception.CustomException;
 import org.unilab.improfessorbe.global.exception.ErrorCode;
