+++ src/main/java/org/unilab/improfessorbe/domain/problem/controller/ProblemController.java
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import org.unilab.improfessorbe.domain.parse.util.FileLogUtil;
-import org.unilab.improfessorbe.domain.problem.application.dto.ProblemDownloadResponse;
-import org.unilab.improfessorbe.domain.problem.application.dto.ProblemGenerationResponse;
+import org.unilab.improfessorbe.domain.problem.dto.ProblemDownloadResponse;
+import org.unilab.improfessorbe.domain.problem.dto.ProblemGenerationResponse;
 //import org.unilab.improfessorbe.domain.problem.application.dto.ProblemResponse;
-import org.unilab.improfessorbe.domain.problem.application.service.ProblemService;
+import org.unilab.improfessorbe.domain.problem.service.ProblemService;
 import org.unilab.improfessorbe.domain.user.service.UserService;
 import org.unilab.improfessorbe.global.common.ApiResponse;
 import org.unilab.improfessorbe.global.exception.CustomException;
