+++ src/main/java/org/unilab/improfessorbe/domain/problem/dto/ProblemDownloadResponse.java
@@ -1,4 +1,4 @@
-package org.unilab.improfessorbe.domain.problem.application.dto;
+package org.unilab.improfessorbe.domain.problem.dto;
 
 import lombok.Builder;
 import lombok.Data;
