+++ src/main/java/org/unilab/improfessorbe/domain/user/service/UserRecommendationService.java
@@ -2,7 +2,7 @@ package org.unilab.improfessorbe.domain.user.service;
 
 import org.springframework.stereotype.Service;
 import org.unilab.improfessorbe.domain.user.domain.User;
-import org.unilab.improfessorbe.domain.user.repository.UserRepository;
+import org.unilab.improfessorbe.domain.user.infrastructure.repository.UserRepository;
 import org.unilab.improfessorbe.global.exception.CustomException;
 import org.unilab.improfessorbe.global.exception.ErrorCode;
