+++ src/main/java/org/unilab/improfessorbe/domain/user/service/UserService.java
@@ -18,7 +18,7 @@ import org.unilab.improfessorbe.domain.user.dto.request.UserRegisterRequest;
 import org.unilab.improfessorbe.domain.user.dto.request.UserUpdateRequest;
 import org.unilab.improfessorbe.domain.user.dto.response.UserLoginResponse;
 import org.unilab.improfessorbe.domain.user.dto.response.UserResponse;
-import org.unilab.improfessorbe.domain.user.repository.UserRepository;
+import org.unilab.improfessorbe.domain.user.infrastructure.repository.UserRepository;
 import org.unilab.improfessorbe.global.exception.CustomException;
 import org.unilab.improfessorbe.global.exception.ErrorCode;
 import org.unilab.improfessorbe.global.security.jwt.JwtToken;
